# docs/openapi.yaml
openapi: "3.0.3"

info:
  title: gogocoin API
  version: "1.0.0"
  description: |
    gogocoin の REST API。取引制御・状態確認・設定管理・データ参照を提供する。
    すべてのエンドポイントはデフォルトで `http://localhost:8080` で提供される。

servers:
  - url: http://localhost:8080
    description: ローカル（Docker）

tags:
  - name: system
    description: ヘルスチェック・システム状態
  - name: data
    description: 残高・取引履歴・注文・パフォーマンス・ログ
  - name: trading
    description: 取引の開始・停止・状態確認
  - name: config
    description: 設定の取得・更新
  - name: strategy
    description: 戦略管理

paths:
  /health:
    get:
      operationId: GetHealth
      tags: [system]
      summary: ヘルスチェック
      description: サーバーの生存確認。ロードバランサーや監視ツール向け。
      responses:
        "200":
          description: サーバー正常稼働中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/status:
    get:
      operationId: GetApiStatus
      tags: [system]
      summary: システム状態の取得
      description: 稼働時間・取引統計・現在価格・シグナル等を返す。
      parameters:
        - name: symbol
          in: query
          description: 価格取得対象のシンボル（省略時は設定の第一シンボル）
          required: false
          schema:
            type: string
            example: XRP_JPY
      responses:
        "200":
          description: システム状態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

  /api/balance:
    get:
      operationId: GetApiBalance
      tags: [data]
      summary: 残高取得
      description: |
        bitFlyer から現在の残高を取得する。レート制限時はDBキャッシュにフォールバックする。
        通貨の表示順: JPY → BTC → ETH → XRP → XLM → MONA → その他（アルファベット順）
      responses:
        "200":
          description: 残高一覧（空の場合は空配列）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/trades:
    get:
      operationId: GetApiTrades
      tags: [data]
      summary: 取引履歴の取得
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - name: since
          in: query
          description: |
            取得範囲を時刻で絞り込む。
            - `today`: JST 当日 0:00 以降の全トレード（limit は無視）
            - RFC3339 タイムスタンプ (例: `2026-04-19T00:00:00+09:00`): その時刻以降のトレード
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 取引履歴（新しい順）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/orders:
    get:
      operationId: GetApiOrders
      tags: [data]
      summary: 注文履歴の取得
      description: 取引履歴から order_id を持つレコードを注文として返す。
      parameters:
        - name: limit
          in: query
          description: "取得件数（デフォルト: 20、最大: 100）"
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: 注文一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/performance:
    get:
      operationId: GetApiPerformance
      tags: [data]
      summary: パフォーマンス指標の取得
      description: 過去30日分の取引パフォーマンス統計を返す。
      responses:
        "200":
          description: パフォーマンス指標一覧（新しい順）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformanceMetric"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/logs:
    get:
      operationId: GetApiLogs
      tags: [data]
      summary: ログの取得
      parameters:
        - name: limit
          in: query
          description: "取得件数（デフォルト: 100、最大: 1000）"
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: level
          in: query
          description: ログレベルフィルター
          required: false
          schema:
            type: string
            enum: [DEBUG, INFO, WARN, ERROR]
        - name: category
          in: query
          description: カテゴリフィルター
          required: false
          schema:
            type: string
            enum: [system, api, strategy, trading, data, ui]
      responses:
        "200":
          description: ログ一覧（新しい順）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/config:
    get:
      operationId: GetApiConfig
      tags: [config]
      summary: 現在の設定を取得
      description: メモリ上の設定オブジェクト全体を返す。APIキー等の機密情報はプレースホルダーで返される。
      responses:
        "200":
          description: 現在の設定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
    post:
      operationId: PostApiConfig
      tags: [config]
      summary: 設定を更新
      description: |
        戦略名またはリスクパラメータを更新し `configs/config.yaml` に保存する。
        APIキー等の機密情報は更新できない（実行時変更不可）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigUpdateRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigUpdateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/strategy/reset:
    post:
      operationId: PostApiStrategyReset
      tags: [strategy]
      summary: 戦略状態のリセット
      description: 現在の戦略の内部状態（EMAキャッシュ・クールダウン等）を初期化する。
      responses:
        "200":
          description: リセット成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleSuccess"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /api/trading/start:
    post:
      operationId: PostApiTradingStart
      tags: [trading]
      summary: 取引開始
      description: 取引フラグを有効化し、DB に状態を永続化する。
      responses:
        "200":
          description: 取引開始成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingStatusResponse"
        "500":
          description: 取引開始失敗
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingStatusResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /api/trading/stop:
    post:
      operationId: PostApiTradingStop
      tags: [trading]
      summary: 取引停止
      description: 取引フラグを無効化し、DB に状態を永続化する。進行中の注文はキャンセルしない。
      responses:
        "200":
          description: 取引停止成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingStatusResponse"
        "500":
          description: 取引停止失敗
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingStatusResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /api/trading/status:
    get:
      operationId: GetApiTradingStatus
      tags: [trading]
      summary: 取引状態の取得
      responses:
        "200":
          description: 現在の取引状態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingStatusResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

components:
  parameters:
    LimitParam:
      name: limit
      in: query
      description: "取得件数（デフォルト: 50）"
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1

  responses:
    BadRequest:
      description: パラメータ不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: アプリケーション未初期化
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Internal server error"

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        timestamp:
          type: string
          format: date-time
        uptime:
          type: string
          example: "1d 3h 22m"
        version:
          type: string
          example: v1.0.0

    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: running
        credentials_status:
          type: string
          example: trading
        strategy:
          type: string
          example: scalping
        last_update:
          type: string
          format: date-time
        uptime:
          type: string
          example: "2h 15m"
        total_trades:
          type: integer
        active_orders:
          type: integer
        trading_enabled:
          type: boolean
        monitoring_symbols:
          type: array
          items:
            type: string
          example: ["XRP_JPY"]
        monitoring_prices:
          type: object
          additionalProperties:
            type: number
          example: {"XRP_JPY": 95.23}
        current_price:
          type: number
          example: 95.23
        current_signal:
          type: string
          example: "N/A"

    Balance:
      type: object
      properties:
        id:
          type: integer
        currency:
          type: string
          example: JPY
        amount:
          type: number
          example: 10000.0
        available:
          type: number
          example: 9500.0
        timestamp:
          type: string
          format: date-time

    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
          example: XRP_JPY
        product_code:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          example: MARKET
        amount:
          type: number
        size:
          type: number
          example: 100.0
        price:
          type: number
          example: 95.23
        fee:
          type: number
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELED, EXPIRED, REJECTED]
        order_id:
          type: string
        executed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        strategy_name:
          type: string
          example: scalping
        pnl:
          type: number
          example: 12.5

    OrderResponse:
      type: object
      properties:
        order_id:
          type: string
        symbol:
          type: string
          example: XRP_JPY
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          example: MARKET
        size:
          type: number
        price:
          type: number
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELED, EXPIRED, REJECTED]
        executed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    PerformanceMetric:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date-time
        total_return:
          type: number
        daily_return:
          type: number
        win_rate:
          type: number
          description: 勝率（0.0 〜 1.0）
        max_drawdown:
          type: number
        sharpe_ratio:
          type: number
        profit_factor:
          type: number
        total_trades:
          type: integer
        winning_trades:
          type: integer
        losing_trades:
          type: integer
        average_win:
          type: number
        average_loss:
          type: number
        largest_win:
          type: number
        largest_loss:
          type: number
        consecutive_wins:
          type: integer
        consecutive_loss:
          type: integer
        total_pnl:
          type: number

    LogEntry:
      type: object
      properties:
        level:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR]
        category:
          type: string
          enum: [system, api, strategy, trading, data, ui]
        message:
          type: string
        fields:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time

    Config:
      type: object
      description: アプリケーション設定全体。詳細は docs/CONFIG.md を参照。
      properties:
        trading:
          type: object
          properties:
            strategy:
              type: object
              properties:
                name:
                  type: string
                  example: scalping
        strategy_params:
          type: object

    ConfigUpdateRequest:
      type: object
      properties:
        strategy:
          type: object
          properties:
            name:
              type: string
              enum: [scalping, simple]
              example: scalping
        risk:
          type: object
          properties:
            stop_loss:
              type: number
              minimum: 0
              maximum: 100
              description: 損切パーセント（0〜100）
            take_profit:
              type: number
              minimum: 0
              maximum: 100
              description: 利確パーセント（0〜100）

    ConfigUpdateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Configuration saved
        config:
          $ref: "#/components/schemas/Config"

    TradingStatusResponse:
      type: object
      properties:
        enabled:
          type: boolean
        status:
          type: string
          enum: [running, stopped, success, error]
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    SimpleSuccess:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Strategy reset successfully